# alphaforge-brain/src/api/routes/run_events.py
from __future__ import annotations

import json
from collections.abc import AsyncGenerator
from datetime import datetime, timezone
from typing import Any

from domain.errors import NotFoundError  # noqa: F401  (for future use)
from domain.run.event_buffer import GLOBAL_EVENT_BUFFERS, EventRingBuffer
from fastapi import APIRouter, HTTPException, Request, Response
from fastapi.responses import StreamingResponse

"""Server-Sent Events stream for run progress (T051).

Baseline implementation (synchronous orchestrator completed immediately) emits:
1. An initial heartbeat event (type=heartbeat)
2. A single snapshot event (type=snapshot) with summary + validation p_values if run exists
3. Periodic heartbeats every `HEARTBEAT_SEC` until generator ends (currently ends immediately after snapshot)

Forward plan (T052+):
- Introduce ring buffer and keep connection open while async orchestration progresses.
- Support Last-Event-ID header for resume using in-memory ring buffer.
"""

router = APIRouter()

HEARTBEAT_SEC = 15  # placeholder; not really used while we end stream early


def _event(event_id: int, event_type: str, data: dict[str, Any]) -> str:
    payload = json.dumps(
        {"type": event_type, "ts": datetime.now(timezone.utc).isoformat(), "data": data}
    )
    return f"id: {event_id}\nevent: {event_type}\ndata: {payload}\n\n"


async def _stream(
    request: Request, run_hash: str, last_event_id: int | None
) -> AsyncGenerator[bytes, None]:
    # Access registry from app closure by attribute introspection (app factory stores it inside route closures only).
    # Because registry isn't exposed, we re-fetch run detail via internal app client logic by calling the GET handler.
    # Simpler: reach into app.routes to find get_run dependency - but for baseline we replicate minimal logic from app.
    app = request.app  # already created

    # Reconstruct the registry by reading the list_runs handler closure cell.
    # Simpler approach: store registry on app.state when created (future refactor). For now we attempt attribute injection.
    registry = getattr(app.state, "registry", None)
    if registry is None:
        # Attempt to locate via function closure (fragile but acceptable baseline). Fallback: error.
        for route in app.routes:
            if getattr(route, "path", "").startswith("/runs/{run_hash}") and hasattr(
                route, "endpoint"
            ):
                fn = route.endpoint
                if getattr(fn, "__name__", "") == "get_run" and getattr(
                    fn, "__closure__", None
                ):
                    for cell in fn.__closure__:
                        if hasattr(cell.cell_contents, "store"):
                            registry = cell.cell_contents
                            break
            if registry is not None:
                break
    if registry is None:
        raise HTTPException(status_code=500, detail="registry unavailable")

    # Per-run ring buffer container on app
    if not hasattr(app.state, "event_buffers"):
        app.state.event_buffers = {}
    bufs: dict[str, EventRingBuffer] = app.state.event_buffers
    if run_hash not in bufs:
        if run_hash in GLOBAL_EVENT_BUFFERS:
            bufs[run_hash] = GLOBAL_EVENT_BUFFERS[run_hash]
        else:
            bufs[run_hash] = EventRingBuffer(capacity=256)

    rec = registry.get(run_hash)
    if not rec:
        raise HTTPException(status_code=404, detail="run not found")

    # For synchronous completed runs we provide a deterministic compact event sequence independent of internal buffer:
    # id:0 heartbeat, id:1 snapshot. Subsequent calls with Last-Event-ID>=1 yield nothing unless a cancellation occurs
    # (which appends a real cancellation event we expose with id:2). This matches tests expecting empty resume.
    status = rec.get("status", "COMPLETE")
    if status == "COMPLETE":
        # Cancellation events may have been appended later; detect if buffer has a cancellation and expose it as id 2
        buffer = bufs[run_hash]
        cancelled_events = [e for e in buffer.since(None) if e.type == "cancelled"]
        if last_event_id is None:
            # Fresh client -> send heartbeat + snapshot (ids 0 & 1)
            yield _event(0, "heartbeat", {"status": status}).encode()
            validation_v2 = rec.get("validation_v2")
            if not isinstance(validation_v2, dict):
                validation_v2 = None
            snapshot = {
                "run_hash": run_hash,
                "summary": rec.get("summary"),
                "p_values": rec.get("p_values"),
                # Surface validation summary for parity with /runs/{run_hash} detail endpoint
                # (legacy alias 'validation' retained for backward compatibility with older clients/tests)
                "validation_summary": rec.get("validation_summary"),
                "validation": rec.get("validation_summary"),
                "validation_schema_version": rec.get("validation_schema_version")
                or (validation_v2.get("schema_version") if validation_v2 else None),
                "validation_manifest": rec.get("validation_manifest")
                or (validation_v2.get("manifest") if validation_v2 else None),
                "validation_manifest_hash": rec.get("validation_manifest_hash")
                or (validation_v2.get("manifest_hash") if validation_v2 else None),
                "validation_significance": rec.get("validation_significance")
                or (
                    validation_v2.get("significance_status") if validation_v2 else None
                ),
                "validation_modules": (
                    validation_v2.get("modules") if validation_v2 else None
                ),
                "validation_config": (
                    validation_v2.get("config") if validation_v2 else None
                ),
                "validation_sections": {
                    "permutation": (
                        validation_v2.get("permutation") if validation_v2 else None
                    ),
                    "bias_adjustments": (
                        validation_v2.get("bias_adjustments") if validation_v2 else None
                    ),
                    "cross_validation": (
                        validation_v2.get("cross_validation") if validation_v2 else None
                    ),
                    "execution_realism": (
                        validation_v2.get("execution_realism")
                        if validation_v2
                        else None
                    ),
                },
                "validation_artifacts": (
                    validation_v2.get("artifacts") if validation_v2 else None
                ),
                "validation_correlation_id": f"{run_hash}:validation_snapshot",
                "status": status,
            }
            yield _event(1, "snapshot", snapshot).encode()
        elif last_event_id == 0:
            # Client has heartbeat only, send snapshot
            validation_v2 = rec.get("validation_v2")
            if not isinstance(validation_v2, dict):
                validation_v2 = None
            snapshot = {
                "run_hash": run_hash,
                "summary": rec.get("summary"),
                "p_values": rec.get("p_values"),
                "validation_summary": rec.get("validation_summary"),
                "validation": rec.get("validation_summary"),
                "validation_schema_version": rec.get("validation_schema_version")
                or (validation_v2.get("schema_version") if validation_v2 else None),
                "validation_manifest": rec.get("validation_manifest")
                or (validation_v2.get("manifest") if validation_v2 else None),
                "validation_manifest_hash": rec.get("validation_manifest_hash")
                or (validation_v2.get("manifest_hash") if validation_v2 else None),
                "validation_significance": rec.get("validation_significance")
                or (
                    validation_v2.get("significance_status") if validation_v2 else None
                ),
                "validation_modules": (
                    validation_v2.get("modules") if validation_v2 else None
                ),
                "validation_config": (
                    validation_v2.get("config") if validation_v2 else None
                ),
                "validation_sections": {
                    "permutation": (
                        validation_v2.get("permutation") if validation_v2 else None
                    ),
                    "bias_adjustments": (
                        validation_v2.get("bias_adjustments") if validation_v2 else None
                    ),
                    "cross_validation": (
                        validation_v2.get("cross_validation") if validation_v2 else None
                    ),
                    "execution_realism": (
                        validation_v2.get("execution_realism")
                        if validation_v2
                        else None
                    ),
                },
                "validation_artifacts": (
                    validation_v2.get("artifacts") if validation_v2 else None
                ),
                "validation_correlation_id": f"{run_hash}:validation_snapshot",
                "status": status,
            }
            yield _event(1, "snapshot", snapshot).encode()
        elif last_event_id >= 1 and cancelled_events:
            # If a cancellation event occurred after completion, expose it as synthetic id 2
            if last_event_id < 2:
                yield _event(
                    2, "cancelled", {"run_hash": run_hash, "status": "CANCELLED"}
                ).encode()
        return

    # Fallback (future async path) - just stream raw buffer events
    buffer = bufs[run_hash]
    to_send = buffer.since(last_event_id)
    for ev in to_send:
        yield _event(ev.id, ev.type, ev.data).encode()
    return


@router.get("/runs/{run_hash}/events", tags=["runs"], include_in_schema=True)
async def run_events(
    request: Request, run_hash: str, after_id: int | None = None
) -> Response:
    """Flush current buffered events (legacy one-shot) with optional incremental filter.

    Enhancements (T079):
    - Query param `after_id` returns only events with id > after_id.
    - ETag header (ETag: "run_hash:last_id") allowing client caching. If client sends If-None-Match
      matching current ETag and no new events beyond after_id -> 304.
    """
    # Reuse underlying logic by materializing generator results immediately (short runs only)
    lei_hdr = request.headers.get("last-event-id") or request.headers.get(
        "Last-Event-ID"
    )
    last_event_id = after_id
    if lei_hdr and last_event_id is None:  # backward compat with old tests
        try:
            parsed = int(lei_hdr)
            if parsed >= 0:
                last_event_id = parsed
        except ValueError:
            last_event_id = None

    # Materialize events
    chunks: list[bytes] = []
    async for part in _stream(request, run_hash, last_event_id):
        chunks.append(part)
    body = b"".join(chunks)

    # Compute last id if any
    last_id = None
    if chunks:
        # parse last id from last chunk (first line like 'id: N')
        for line in chunks[-1].decode().splitlines():
            if line.startswith("id: "):
                try:
                    last_id = int(line[4:].strip())
                except ValueError:
                    pass
                break
    etag_val = f'{run_hash}:{last_id if last_id is not None else "-"}'
    inm = request.headers.get("if-none-match") or request.headers.get("If-None-Match")
    if inm == etag_val and not body:
        return Response(status_code=304)
    resp = Response(content=body, media_type="text/event-stream")
    resp.headers["ETag"] = etag_val
    return resp


async def _heartbeat_stream(
    interval_sec: int, request: Request
) -> AsyncGenerator[bytes, None]:  # pragma: no cover (timing logic)
    import asyncio

    while True:
        if await request.is_disconnected():
            break
        yield _event(-1, "heartbeat", {"keepalive": True}).encode()
        await asyncio.sleep(interval_sec)


@router.get("/runs/{run_hash}/events/stream", tags=["runs"], include_in_schema=True)
async def run_events_stream(request: Request, run_hash: str) -> StreamingResponse:
    """Long-lived incremental SSE (T078).

    Behavior:
    - Reads Last-Event-ID header if present for resume.
    - Streams any current buffered events, then keeps connection open.
    - Emits heartbeat every HEARTBEAT_SEC while waiting for new events.
    - Stops when run reaches terminal state and client has received final snapshot/completed event.
    """
    lei_hdr = request.headers.get("last-event-id") or request.headers.get(
        "Last-Event-ID"
    )
    last_event_id: int | None = None
    if lei_hdr:
        try:
            last_event_id = int(lei_hdr)
        except ValueError:
            last_event_id = None

    import asyncio

    async def generator() -> AsyncGenerator[bytes, None]:  # incremental
        nonlocal last_event_id
        app = request.app
        if not hasattr(app.state, "event_buffers"):
            app.state.event_buffers = {}
        # Obtain registry similarly to _stream
        registry = getattr(app.state, "registry", None)
        if registry is None:
            raise HTTPException(status_code=500, detail="registry unavailable")
        bufs: dict[str, EventRingBuffer] = app.state.event_buffers
        if run_hash not in bufs:
            if run_hash in GLOBAL_EVENT_BUFFERS:
                bufs[run_hash] = GLOBAL_EVENT_BUFFERS[run_hash]
            else:
                bufs[run_hash] = EventRingBuffer(capacity=256)
        buffer = bufs[run_hash]
        # On initial connect, if no last_event_id, emit buffer contents
        initial = buffer.since(last_event_id)
        for ev in initial:
            yield _event(ev.id, ev.type, ev.data).encode()
            last_event_id = ev.id

        # Heartbeat + wait loop
        while True:
            if await request.is_disconnected():
                break
            rec = registry.get(run_hash)
            if rec is None:
                # run not found -> single error event then stop
                yield _event(
                    (last_event_id or 0) + 1, "error", {"code": "NOT_FOUND"}
                ).encode()
                break
            # New events?
            new_events = buffer.since(last_event_id)
            terminal = rec.get("status") in {"COMPLETE", "FAILED", "CANCELLED"}
            # Fast-path: synchronous orchestrator may populate summary before status flips to COMPLETE.
            # If we have a summary and no new buffered events, treat as terminal to avoid long heartbeat sleep (test hang).
            if (not terminal) and (rec.get("summary") is not None) and (not new_events):
                terminal = True  # treat as completed for streaming purposes
            if new_events:
                for ev in new_events:
                    yield _event(ev.id, ev.type, ev.data).encode()
                    last_event_id = ev.id
            else:
                # heartbeat
                yield _event(
                    (last_event_id or 0), "heartbeat", {"status": rec.get("status")}
                ).encode()
            if terminal and (not buffer.since(last_event_id)):
                break
            # Reduce sleep while run not yet terminal to keep test latency low; once terminal reached we exit above.
            await asyncio.sleep(0.05 if not terminal else HEARTBEAT_SEC)

    return StreamingResponse(generator(), media_type="text/event-stream")