# specs/001-initial-dual-tier/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Project A Backend
  version: 0.3.0
paths:
...
  /runs/{run_hash}/events:
    get:
      tags:
      - runs
      summary: Run Events
      description: "Flush current buffered events (legacy one-shot) with optional\
        \ incremental filter.\n\nEnhancements (T079):\n- Query param `after_id` returns\
        \ only events with id > after_id.\n- ETag header (ETag: \"run_hash:last_id\"\
        ) allowing client caching. If client sends If-None-Match\n  matching current\
        \ ETag and no new events beyond after_id -> 304."
      operationId: run_events_runs__run_hash__events_get
      parameters:
      - name: run_hash
        in: path
        required: true
        schema:
          type: string
          title: Run Hash
      - name: after_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: After Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /runs/{run_hash}/events/stream:
    get:
      tags:
      - runs
      summary: Run Events Stream
      description: 'Long-lived incremental SSE (T078).


        Behavior:

        - Reads Last-Event-ID header if present for resume.

        - Streams any current buffered events, then keeps connection open.

        - Emits heartbeat every HEARTBEAT_SEC while waiting for new events.

        - Stops when run reaches terminal state and client has received final snapshot/completed
        event.'
      operationId: run_events_stream_runs__run_hash__events_stream_get
      parameters:
      - name: run_hash
        in: path
        required: true
        schema:
          type: string
          title: Run Hash
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /runs/{run_hash}:
    get:
      tags:
      - runs
      summary: Get Run Detail
      operationId: get_run_detail_runs__run_hash__get
      parameters:
      - name: run_hash
        in: path
        required: true
        schema:
          type: string
          title: Run Hash
      - name: include_anomalies
        in: query
        required: false
        schema:
          type: boolean
          description: Include dataset anomaly counters when available
          default: false
          title: Include Anomalies
        description: Include dataset anomaly counters when available
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /runs/{run_hash}/artifacts:
    get:
      tags:
      - runs
      summary: List Artifacts
      operationId: list_artifacts_runs__run_hash__artifacts_get
      parameters:
      - name: run_hash
        in: path
        required: true
        schema:
          type: string
          title: Run Hash
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response List Artifacts Runs  Run Hash  Artifacts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ArtifactDescriptor:
      properties:
        name:
          type: string
          title: Name
        sha256:
          type: string
          title: Sha256
        size:
          type: integer
          title: Size
      type: object
      required:
      - name
      - sha256
      - size
      title: ArtifactDescriptor
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RunDetailResponse:
      properties:
        run_id:
          type: string
          title: Run Id
        run_hash:
          type: string
          title: Run Hash
        status:
          type: string
          title: Status
        phase:
          type: string
          title: Phase
        artifacts:
          items:
            $ref: '#/components/schemas/ArtifactDescriptor'
          type: array
          title: Artifacts
        summary:
          anyOf:
          - type: object
          - type: 'null'
          title: Summary
        data_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Data Hash
        calendar_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Calendar Id
        validation_summary:
          anyOf:
          - type: object
          - type: 'null'
          title: Validation Summary
        validation:
          anyOf:
          - type: object
          - type: 'null'
          title: Validation
        manifest:
          anyOf:
          - type: object
          - type: 'null'
          title: Manifest
        api_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Api Version
        schema_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Schema Version
        content_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Hash
        pinned:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Pinned
        retention_state:
          anyOf:
          - type: string
          - type: 'null'
          title: Retention State
        metrics_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Metrics Hash
        equity_curve_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Equity Curve Hash
        validation_schema_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Validation Schema Version
        validation_manifest:
          anyOf:
          - type: object
          - type: 'null'
          title: Validation Manifest
        validation_significance:
          anyOf:
          - type: string
          - type: 'null'
          title: Validation Significance
        validation_manifest_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Validation Manifest Hash
      type: object
      required:
      - run_id
      - run_hash
      - status
      - phase
      - artifacts
      title: RunDetailResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...